# backend/RefugeeShelter/src/main/resources/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "refugee-shelter-api",
    "description": "The API for the Refugee Shelter project",
    "contact": {
      "email": "akrikoff@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://virtserver.swaggerhub.com/AKRIKOFF/RefugeeShelter/1.0.0",
      "description": "SwaggerHub API Auto Mocking"
    }
  ],
  "paths": {
...
    "/api/v1/reservations": {
      "get": {
        "tags": [
          "Reservation"
        ],
        "description": "Returns all reservations",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization Bearer",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "All reservations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/inline_response_200_3"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_403"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/reservations/{reservationId}": {
      "get": {
        "tags": [
          "Reservation"
        ],
        "description": "Return reservation by id",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization Bearer",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reservationId",
            "in": "path",
            "description": "Reservation ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reservation by ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_3"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_403"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Reservation"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization Bearer",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reservationId",
            "in": "path",
            "description": "Reservation ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/reservations_reservationId_body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated Reservation by ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_4"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_403"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Reservation"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization Bearer",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "reservationId",
            "in": "path",
            "description": "Reservation ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete Reservation by ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_4"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_403"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/reservations/users/{userId}": {
      "get": {
        "tags": [
          "Reservation"
        ],
        "description": "Return reservation by owner id",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization Bearer",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "Owner ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reservation by owner ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/inline_response_200_3"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_403"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/reservations/users/{userId}/rooms/{roomId}": {
      "post": {
        "tags": [
          "Reservation"
        ],
        "description": "Create reservation",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Authorization Bearer",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "User ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "roomId",
            "in": "path",
            "description": "Room ID",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/rooms_roomId_body_2"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reservation by owner ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/inline_response_200_3"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_403"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Room": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "address": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "price": {
            "type": "integer"
          },
          "hasKitchen": {
            "type": "boolean"
          },
          "hasBathroom": {
            "type": "boolean"
          },
          "description": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          }
        }
      },
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "inline_response_403": {
        "type": "object",
        "properties": {
          "error_message": {
            "type": "string",
            "example": "The user is not logged in"
          }
        }
      },
...
      "inline_response_200_3": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 64
          },
          "startDate": {
            "type": "string",
            "example": "2022-01-01"
          },
          "endDate": {
            "type": "string",
            "example": "2022-01-01"
          },
          "room": {
            "$ref": "#/components/schemas/Room"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "reservations_reservationId_body": {
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string",
            "example": "2022-03-01"
          },
          "endDate": {
            "type": "string",
            "example": "2022-04-01"
          }
        }
      },
      "inline_response_200_4": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 64
          },
          "startDate": {
            "type": "string",
            "example": "2022-03-01"
          },
          "endDate": {
            "type": "string",
            "example": "2022-04-01"
          },
          "room": {
            "$ref": "#/components/schemas/Room"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          }
        }
      },
      "rooms_roomId_body_2": {
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string",
            "example": "2022-01-01"
          },
          "endDate": {
            "type": "string",
            "example": "2022-01-01"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}